refactor(core): move material design icons into the material design loader
