+++ packages/backbrace-core/src/providers/icons.js
@@ -1,33 +1,13 @@
 /**
- * Icons provider module. By default, it gets icons from Material Design Icons Webfont
- * (https://materialdesignicons.com/).
+ * Icons provider module.
  * @module iconsprovider
  * @private
  */
 
 /** @type {IconsInstance} */
 let instance = {
-    get: function(name, className) {
-
-        // Set defaults.
-        name = name || 'alert';
-
-        // Map template.
-        if (name.indexOf('%') === 0) {
-            if (name === '%new%') {
-                name = 'plus';
-            } else if (name === '%delete%') {
-                name = 'close';
-            } else {
-                name = name.replace(/%/g, '');
-            }
-        }
-
-        // Prepend mdi- if missing.
-        if (name.indexOf('mdi-') !== 0)
-            name = 'mdi-' + name;
-
-        return '<i class="mdi ' + name + (className ? ' ' + className : '') + '" />';
+    get: function(name) {
+        return name;
     }
 };
